Stop blocking drawing of the page
- During development, I linked to the full version of the Angular
library in order to get friendly debugging messages in the console. When
I uploaded 'index.html' to the server, I noticed very slow page load and
found out through the Network panel of debugger tools that the loading
of the 1.1 MB Angular file was blocking the rest of the page from
loading, so now I'm loading the online version of Angular for much
better page responsiveness.

Linking to the minified version of Angular may make it harder to easily
click on console error messages to find the exact Angular error. Linking
to a 'map' file may alleviate this but I haven't tested it.